Haven't really changed much, but added in new interfaces to try and help me think of how it will be structured. But trying to get a shadermanager that can deal with HLSL and GLSL with their different inputs and access to data could prove difficult, so I need to work on what is their middle ground. Otherwise I think I'll work on VBO until I can think of something.

+++ Project-LimeViolet/scr/Render/DX11Render.cpp
@@ -251,5 +251,5 @@ void DX11Render::UpdateViewMatrix()
 
 void DX11Render::UpdateProjectionMatrix()
 {
-	DirectX::XMStoreFloat4x4(reinterpret_cast<DirectX::XMFLOAT4X4*>(&_projection.m), DirectX::XMMatrixPerspectiveFovLH(_activeCamera->GetFOV(), _windowWidth / _windowHeight, _activeCamera->GetNear(), _activeCamera->GetFar()));
+	DirectX::XMStoreFloat4x4(reinterpret_cast<DirectX::XMFLOAT4X4*>(&_projection.m), DirectX::XMMatrixPerspectiveFovLH(_activeCamera->GetFOV(), static_cast<float>(_windowWidth) / static_cast<float>(_windowHeight), _activeCamera->GetNear(), _activeCamera->GetFar()));
 }
